integrated-cost-optimizer/pyproject.toml
[project]
name = "integrated-cost-optimizer"
version = "1.0.0"
description = "Integrated LLM Cost Optimizer combining Prompt Optimization, Semantic Caching, and Request Batching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "google-generativeai>=0.3.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
